테스트 성공: view hierarchy
